Rename the aperture region conversion module
